moved normalizeColor and changed dependencies
