+++ Libraries/StyleSheet/__tests__/setNormalizedColorAlpha-test.js
@@ -11,7 +11,7 @@
 'use strict';
 
 const setNormalizedColorAlpha = require('../setNormalizedColorAlpha');
-const normalizeColor = require('../../Color/normalizeColor');
+const normalizeColor = require('../normalizeColor');
 
 describe('setNormalizedColorAlpha', function() {
   it('should adjust the alpha of the color passed in', function() {
